fix: address review comments
- Move write-lock acquisition after send_handle.await to avoid
  blocking polling during bulk send
- Add else branch to pollJobStatus to stop polling on failures
- Check job.status === 'failed' before job.failed === 0 in toast
- Add MailJob/MailJobStatus to generate-types binary
- Change MailJob usize fields to i32 for Specta TS compatibility

+++ backend/src/models.rs
@@ -81,9 +81,9 @@ pub enum MailJobStatus {
 pub struct MailJob {
     pub id: String,
     pub status: MailJobStatus,
-    pub total_recipients: usize,
-    pub sent: usize,
-    pub failed: usize,
+    pub total_recipients: i32,
+    pub sent: i32,
+    pub failed: i32,
     pub failed_recipients: Vec<String>,
     pub error: Option<String>,
     pub created_at: chrono::DateTime<chrono::Utc>,
